Problem: tesseract not ocring numeral-only images
A crop of only a page number fed to
tesseract is returning only an empty
string. Adding words to image results
in ocring of words, but not numeral.

Solution: The Page Segmentation Mode (PSM)
used is important for numeral only images.
Using a PSM of 6 results in numerals being
correctly recognized.